Add further options to CodePen component
Properties added are:
clickToLoad - For running the CodePen in a preview mode that is less resource-intensive.
editable - To allow viewers to edit and run the pen on the page
theme - To choose between default, light, and dark themes. `string` was left as a generic type because "Pro" CodePen accounts can create their own themes and name them, so the list is not finite.